Content2Cart_backend/requirements.txt
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
google-generativeai==0.3.1
openai==1.3.5
Pillow==10.0.1
python-multipart==0.0.6
pydantic==2.5.2
firebase-admin>=6.3.0
google-cloud-firestore>=2.15.0
